+++ spec/unit/test_type.rb
@@ -287,48 +287,6 @@ module Bmg
 
     end
 
-    describe 'unwrap' do
-      let(:type){
-        Type::ANY
-          .with_attrlist([:sid, :name, :supplies])
-          .with_keys([[:sid], [:sid, :supplies]])
-      }
-
-      it 'drops attrlist' do
-        got = type.unwrap([:supplies])
-        expect(got.to_attrlist).to be_nil
-      end
-
-      it 'keeps keys that don\'t use the wrapped attribute' do
-        got = type.unwrap([:supplies])
-        expect(got.keys).to eql([[:sid]])
-      end
-
-      it 'keeps predicate if it did not use the unwrapped attribute' do
-        from = type
-          .with_predicate(Predicate.eq(:sid, "S1"))
-        got = from
-          .unwrap([:supplies])
-        expect(got.predicate).to eql(from.predicate)
-      end
-
-      it 'split AND-predicate using the unwrapped attribute' do
-        from = type
-          .with_predicate(Predicate.eq(:sid, "S1") & Predicate.neq(:supplies, nil))
-        got = from
-          .unwrap([:supplies])
-        expect(got.predicate).to eql(Predicate.eq(:sid, "S1"))
-      end
-
-      it 'drops OR-predicate using the unwrapped attribute' do
-        from = type
-          .with_predicate(Predicate.eq(:sid, "S1") | Predicate.neq(:supplies, nil))
-        got = from
-          .unwrap([:supplies])
-        expect(got.predicate).to eql(Predicate.tautology)
-      end
-    end
-
     describe 'transform' do
 
       it 'keeps attrlist' do
@@ -382,6 +340,29 @@ module Bmg
 
     end
 
+    describe 'ungroup' do
+      let(:type){
+        Type::ANY
+          .with_attrlist([:sid, :name, :supplies])
+          .with_keys([[:sid]])
+      }
+
+      it 'drops attrlist' do
+        got = type.ungroup([:supplies])
+        expect(got.to_attrlist).to be_nil
+      end
+
+      it "drop keys" do
+        got = type.ungroup([:supplies])
+        expect(got.keys).to be_nil
+      end
+
+      it 'drops predicate' do
+        got = type.ungroup([:supplies])
+        expect(got.predicate).to eql(Predicate.tautology)
+      end
+    end
+
     describe 'union' do
 
       it 'keeps attrlist when known' do
@@ -401,6 +382,48 @@ module Bmg
         expect(t1.union(t2).keys).to eql([[:sid]])
       end
 
+    end
+
+    describe 'unwrap' do
+      let(:type){
+        Type::ANY
+          .with_attrlist([:sid, :name, :supplies])
+          .with_keys([[:sid], [:sid, :supplies]])
+      }
+
+      it 'drops attrlist' do
+        got = type.unwrap([:supplies])
+        expect(got.to_attrlist).to be_nil
+      end
+
+      it 'keeps keys that don\'t use the wrapped attribute' do
+        got = type.unwrap([:supplies])
+        expect(got.keys).to eql([[:sid]])
+      end
+
+      it 'keeps predicate if it did not use the unwrapped attribute' do
+        from = type
+          .with_predicate(Predicate.eq(:sid, "S1"))
+        got = from
+          .unwrap([:supplies])
+        expect(got.predicate).to eql(from.predicate)
+      end
+
+      it 'split AND-predicate using the unwrapped attribute' do
+        from = type
+          .with_predicate(Predicate.eq(:sid, "S1") & Predicate.neq(:supplies, nil))
+        got = from
+          .unwrap([:supplies])
+        expect(got.predicate).to eql(Predicate.eq(:sid, "S1"))
+      end
+
+      it 'drops OR-predicate using the unwrapped attribute' do
+        from = type
+          .with_predicate(Predicate.eq(:sid, "S1") | Predicate.neq(:supplies, nil))
+        got = from
+          .unwrap([:supplies])
+        expect(got.predicate).to eql(Predicate.tautology)
+      end
     end
 
   end